out << this is printing hex without leading "0x" on windows; use (long)thi instead. With this, cling now starts up on Windows!
git-svn-id: https://root.cern.ch/svn/root/branches/dev@39619 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cling/lib/Interpreter/Interpreter.cpp
@@ -258,7 +258,7 @@ namespace cling {
 
     // Set up the gCling variable - even if we use PCH ('this' is different)
     std::stringstream initializer;
-    initializer << "gCling=(cling::Interpreter*)" << this <<";\n";    
+    initializer << "gCling=(cling::Interpreter*)" << (long)this <<";\n";    
     processLine(initializer.str());
 
     handleFrontendOptions();
